Changed `Size2DCollection` to `Collection2D`
And We did it without breaking existing code! Woo!

This change makes that protocol more generic, so it can be applied to sizes, rectangles, anything which conforms.

Part of this was removing the requirement for conforming types to also conform to `Size2D`, and creating a new conformance requirement: `CartesianMeasurable`. We then changed `Size2D` to require `CartesianMeasurable`, and that enabled this broadening.